+++ src/MeteoWeatherAPI/Services/WeatherCacheService.cs
@@ -3,14 +3,6 @@ using Microsoft.Extensions.Caching.Memory;
 
 namespace MeteoWeatherAPI.Services;
 
-public interface IWeatherCacheService
-{
-    void SaveForecast(WeatherForecastDto dto);
-    WeatherForecastDto GetForecastDto(CacheKey cacheKey);
-
-    void DeleteForecast(CacheKey key);
-}
-
 public class WeatherCacheService : IWeatherCacheService
 {
     private readonly IMemoryCache _memoryCache;
